Move InferDomainResponse, LastActionStatus, ViewInfo.ts to separate files
- Remove internal/components/base/actions.ts (moved all actions to more appropriate files)
