+++ packages/components/src/internal/components/domainproperties/DomainForm.tsx
@@ -20,11 +20,8 @@ import { Col, Form, FormControl, Panel, Row } from 'react-bootstrap';
 
 import { Sticky, StickyContainer } from 'react-sticky';
 
-import { AddEntityButton } from '../buttons/AddEntityButton';
-import { ConfirmModal } from '../base/ConfirmModal';
-import { InferDomainResponse } from '../base/models/model';
-import { FileAttachmentForm } from '../files/FileAttachmentForm';
-import { Alert } from '../base/Alert';
+import { AddEntityButton, ConfirmModal, InferDomainResponse, FileAttachmentForm, Alert } from '../../..';
+
 import { FIELD_EDITOR_TOPIC, helpLinkNode } from '../../util/helpLinks';
 
 import { blurActiveElement } from '../../util/utils';
