+++ packages/components/src/internal/components/files/FilePreviewGrid.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import { List, Map } from 'immutable';
 
-import { Alert } from '../base/Alert';
-import { Grid, GridColumn } from '../base/Grid';
-import { InferDomainResponse } from '../base/models/model';
+import { Alert, Grid, GridColumn, InferDomainResponse } from '../../..';
 
 export interface FileGridPreviewProps {
     previewCount: number;
