+++ packages/components/src/internal/components/gridbar/ViewSelector.tsx
@@ -18,9 +18,9 @@ import { DropdownButton, MenuItem } from 'react-bootstrap';
 import { List } from 'immutable';
 
 import { gridSelectView } from '../../actions';
-import { QueryGridModel, ViewInfo } from '../base/models/model';
+import { QueryGridModel } from '../base/models/model';
 import { generateId } from '../../util/utils';
-import { naturalSort } from '../../..';
+import { naturalSort, ViewInfo } from '../../..';
 
 const emptyList = List<ReactNode>();
